+++ db/migrate/20240531115345_update_student_lesson_summaries_to_version_6.rb
@@ -1,0 +1,5 @@
+class UpdateStudentLessonSummariesToVersion6 < ActiveRecord::Migration[7.1]
+  def change
+    update_view :student_lesson_summaries, version: 6, revert_to_version: 5
+  end
+end